requirements.txt
fastapi
uvicorn[standard]
tensorflow==2.19.0
opencv-python
mtcnn
pillow
numpy
